Improve test quality and extract ErrorBoundary component
Extract ErrorBoundary from App.tsx into its own module so tests import
the real component instead of a duplicated copy.  Add makeCredential
factory helper to useCredentials tests to reduce boilerplate.  Replace
raw document.querySelector in UnlockView test with Testing Library
getAllByRole query.  Add test:coverage npm script.
