+++ cmd/tegata-gui/frontend/src/App.test.tsx
@@ -1,41 +1,7 @@
 import { describe, expect, it, vi, beforeEach } from "vitest"
 import { render, screen } from "@testing-library/react"
 import userEvent from "@testing-library/user-event"
-import { Component } from "react"
-import type { ErrorInfo, ReactNode } from "react"
-
-// Import the ErrorBoundary class directly by re-creating it here since
-// it is not exported as a named export from App.tsx (only the default
-// export wraps it). We test the same pattern against the same API.
-class ErrorBoundary extends Component<
-  { children: ReactNode },
-  { error: Error | null }
-> {
-  state: { error: Error | null } = { error: null }
-
-  static getDerivedStateFromError(error: Error) {
-    return { error }
-  }
-
-  componentDidCatch(error: Error, info: ErrorInfo) {
-    console.error("Unhandled render error:", error, info.componentStack)
-  }
-
-  render() {
-    if (this.state.error) {
-      return (
-        <div>
-          <h1>Something went wrong</h1>
-          <p>{this.state.error.message}</p>
-          <button onClick={() => this.setState({ error: null })}>
-            Try again
-          </button>
-        </div>
-      )
-    }
-    return this.props.children
-  }
-}
+import { ErrorBoundary } from "@/components/shared/ErrorBoundary"
 
 function GoodChild() {
   return <div>Everything is fine</div>
@@ -63,7 +29,6 @@ describe("ErrorBoundary", () => {
   })
 
   it("shows 'Something went wrong' and error message when child throws", () => {
-    // Suppress console.error for the expected error
     const spy = vi.spyOn(console, "error").mockImplementation(() => {})
 
     shouldThrow = true
