+++ cmd/tegata-gui/frontend/src/App.tsx
@@ -1,5 +1,4 @@
-import { Component, useCallback, useEffect, useState } from "react"
-import type { ErrorInfo, ReactNode } from "react"
+import { useCallback, useEffect, useState } from "react"
 import { Header } from "@/components/layout/Header"
 import { Sidebar } from "@/components/layout/Sidebar"
 import { CredentialDetail } from "@/components/credentials/CredentialDetail"
@@ -8,6 +7,7 @@ import { SettingsPanel } from "@/components/settings/SettingsPanel"
 import { UnlockView } from "@/components/vault/UnlockView"
 import { SetupWizard } from "@/components/vault/SetupWizard"
 import { LoadingSpinner } from "@/components/shared/LoadingSpinner"
+import { ErrorBoundary } from "@/components/shared/ErrorBoundary"
 import { useVault } from "@/hooks/useVault"
 import { useCredentials } from "@/hooks/useCredentials"
 import { useIdleTimer } from "@/hooks/useIdleTimer"
@@ -172,39 +172,6 @@ function App() {
       />
     </div>
   )
-}
-
-class ErrorBoundary extends Component<
-  { children: ReactNode },
-  { error: Error | null }
-> {
-  state: { error: Error | null } = { error: null }
-
-  static getDerivedStateFromError(error: Error) {
-    return { error }
-  }
-
-  componentDidCatch(error: Error, info: ErrorInfo) {
-    console.error("Unhandled render error:", error, info.componentStack)
-  }
-
-  render() {
-    if (this.state.error) {
-      return (
-        <div className="flex h-screen flex-col items-center justify-center gap-4 bg-background p-8 text-center">
-          <h1 className="text-xl font-semibold text-destructive">Something went wrong</h1>
-          <p className="max-w-md text-sm text-muted-foreground">{this.state.error.message}</p>
-          <button
-            className="rounded-md bg-primary px-4 py-2 text-sm text-primary-foreground"
-            onClick={() => this.setState({ error: null })}
-          >
-            Try again
-          </button>
-        </div>
-      )
-    }
-    return this.props.children
-  }
 }
 
 export default function AppWithBoundary() {
